Displaying user's courses and rooms

+++ src/main/java/com/Lernado/managers/RoomRepository.java
@@ -4,4 +4,4 @@ package com.Lernado.managers;
 import com.Lernado.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Long> {}
+public interface RoomRepository extends JpaRepository<Room, Integer> {}
